eecrm/requirements.txt
black==22.3.0
django-extensions==3.1.5
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
flake8-django==1.1.1
flake8-html==0.4.2
pep8-naming==0.13.0
pip-chill==1.0.1
psycopg2==2.9.3
python-decouple==3.6
